requirements.txt
Django>=4.2
djangorestframework>=3.14.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.4.1   
mysqlclient>=2.2.0        
python-dotenv>=1.0.1        
rest_framework_simplejwt